revert: changes to Info Box
